Splitted Http method handles
