+++ ShadowApiNet/HttpHandlers/HttpConnectHandler.cs
@@ -1,15 +1,19 @@
 ﻿using Microsoft.AspNetCore.Http;
+using Microsoft.EntityFrameworkCore;
 using ShadowApiNet.Interfaces;
 using System;
+using System.Collections.Generic;
+using System.Reflection;
 using System.Threading.Tasks;
 
 namespace ShadowApiNet.HttpHandlers
 {
     internal class HttpConnectHandler : IHttpMethodHandler
     {
-        public Task<HttpContext> Handle(HttpContext httpContext)
+        internal override Task<HttpContext> Handle(HttpContext httpContext, string rootUriPath, string[] pathNodes, DbContext dbContext, Dictionary<PropertyInfo, Type> dbSets, Dictionary<PropertyInfo, PropertyInfo[]> tablesFields)
         {
-            throw new NotImplementedException();
+            this.SetStatusCode(httpContext.Response, StatusCodes.Status501NotImplemented);
+            return Task.Run(() => httpContext);
         }
     }
 }
